Backend/requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
pypdf2
beautifulsoup4
requests
folium
streamlit-folium
python-dotenv